requirements.txt
fastapi
uvicorn
openai
jinja2
aiofiles
python-dotenv
supabase
